chore: move sidebar components into general components folder, update sidebar components
