--- app/src/components/guides/modals/index.ts
@@ -1,5 +1,0 @@
-export * from "./BaseGuideModal";
-export * from "./ContributorsModal";
-export * from "./ObjectivesModal";
-export * from "./RevisionsModal";
-export * from "./VariantsModal";
