+++ app/src/routes/guides/$slug/$variantSlug/index.tsx
@@ -29,7 +29,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { GuideSidebarActions } from "@/components/guides/GuideSidebarActions";
+import { GuideSidebarActions } from "@/components/sidebar/GuideSidebarActions";
 
 function useVote() {
   const [vote, setVote] = useState<"up" | "down" | null>(null);
