+++ app/src/routes/guides/$slug/index.tsx
@@ -30,7 +30,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Route as GuideWalkthroughRoute } from "@/routes/guides/$slug/walkthrough";
-import { GuideSidebarActions } from "@/components/guides/GuideSidebarActions";
+import { GuideSidebarActions } from "@/components/sidebar/GuideSidebarActions";
 
 function useVote() {
   const [vote, setVote] = useState<"up" | "down" | null>(null);
